Update How-to-create-an-optimally-fast-calculated-attribute.md
The updated documentation now includes a fourth way to optimize calculated attributes. By adding the TOP and ORDERBY operators in the SELECT clause, we can further optimize the speed of our calculations. TOP lets us return only the first N elements of the result set, which can be useful if we only need a subset of the data. ORDERBY lets us sort the result set based on a specified column or expression, which can make the data easier and faster to process. Using these operators in the SELECT clause can significantly improve the performance of our calculated attributes.